Initial Test Code
These are all the files needed to run the first json parsing  tests. We
have concluded that we will use the Inklewriter Unity plugin. I will
remove all the unnecesary files in the next commit on this branch. This
will ensure that all test code which I have created so far is backed up.
The next commit will clean all of this up